Fix review comments: add BaseTest and Config
