+++ src/main/java/config/Config.java
@@ -1,0 +1,5 @@
+package config;
+
+public class Config {
+    public final static String URL = "https://qa-scooter.praktikum-services.ru";
+}
